requirements.txt
# Database Connections
psycopg2-binary>=2.9.9
duckdb>=0.9.2
ibis-framework>=7.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Configuration Management
pydantic>=2.5.0
python-dotenv>=1.0.0

# Type Hints
typing-extensions>=4.8.0

# Testing (Optional but recommended)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools (Optional)
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0